feat: ui분리
