+++ src/features/ui/PostTable.tsx
@@ -1,0 +1,105 @@
+import { MessageSquare, Edit2, Trash2, ThumbsUp, ThumbsDown } from "lucide-react"
+import { useAtom } from "jotai"
+import { searchQueryAtom, selectedPostAtom, selectedTagAtom, showEditDialogAtom } from "../../app/atom"
+import useManagePosts from "../useManagePosts"
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "../../shared/ui/table"
+import { HighlightText } from "../highlightText"
+import { Button } from "../../shared/ui/Button"
+import { usePostDetail, useUserModal } from "../useDetail"
+import usePosts from "../usePosts"
+
+type PostTableProps = {
+  updateURL: () => void
+}
+
+const PostTable = ({ updateURL }: PostTableProps) => {
+  const [, setSelectedPost] = useAtom(selectedPostAtom)
+  const { deletePost } = useManagePosts()
+  //   const { likeComment } = useLikeComment()
+  const [, setShowEditDialog] = useAtom(showEditDialogAtom)
+  const { openPostDetail } = usePostDetail()
+  const { posts } = usePosts()
+  const { openUserModal } = useUserModal()
+  const [searchQuery] = useAtom(searchQueryAtom)
+  const [selectedTag, setSelectedTag] = useAtom(selectedTagAtom)
+
+  return (
+    <Table>
+      <TableHeader>
+        <TableRow>
+          <TableHead className="w-[50px]">ID</TableHead>
+          <TableHead>제목</TableHead>
+          <TableHead className="w-[150px]">작성자</TableHead>
+          <TableHead className="w-[150px]">반응</TableHead>
+          <TableHead className="w-[150px]">작업</TableHead>
+        </TableRow>
+      </TableHeader>
+      <TableBody>
+        {posts.map((post) => (
+          <TableRow key={post.id}>
+            <TableCell>{post.id}</TableCell>
+            <TableCell>
+              <div className="space-y-1">
+                <div>{HighlightText(post.title, searchQuery)}</div>
+
+                <div className="flex flex-wrap gap-1">
+                  {post.tags.map((tag) => (
+                    <span
+                      key={tag}
+                      className={`px-1 text-[9px] font-semibold rounded-[4px] cursor-pointer ${
+                        selectedTag === tag
+                          ? "text-white bg-blue-500 hover:bg-blue-600"
+                          : "text-blue-800 bg-blue-100 hover:bg-blue-200"
+                      }`}
+                      onClick={() => {
+                        setSelectedTag(tag)
+                        updateURL()
+                      }}
+                    >
+                      {tag}
+                    </span>
+                  ))}
+                </div>
+              </div>
+            </TableCell>
+            <TableCell>
+              <div className="flex items-center space-x-2 cursor-pointer" onClick={() => openUserModal(post.author)}>
+                <img src={post.author?.image} alt={post.author?.username} className="w-8 h-8 rounded-full" />
+                <span>{post.author?.username}</span>
+              </div>
+            </TableCell>
+            <TableCell>
+              <div className="flex items-center gap-2">
+                <ThumbsUp className="w-4 h-4" />
+                <span>{post.reactions?.likes || 0}</span>
+                <ThumbsDown className="w-4 h-4" />
+                <span>{post.reactions?.dislikes || 0}</span>
+              </div>
+            </TableCell>
+            <TableCell>
+              <div className="flex items-center gap-2">
+                <Button variant="ghost" size="sm" onClick={() => openPostDetail(post)}>
+                  <MessageSquare className="w-4 h-4" />
+                </Button>
+                <Button
+                  variant="ghost"
+                  size="sm"
+                  onClick={() => {
+                    setSelectedPost(post)
+                    setShowEditDialog(true)
+                  }}
+                >
+                  <Edit2 className="w-4 h-4" />
+                </Button>
+                <Button variant="ghost" size="sm" onClick={() => deletePost(post.id)}>
+                  <Trash2 className="w-4 h-4" />
+                </Button>
+              </div>
+            </TableCell>
+          </TableRow>
+        ))}
+      </TableBody>
+    </Table>
+  )
+}
+export default PostTable
